refactor(user): move user to the lib common
