+++ src/main/java/fr/avenirsesr/portfolio/ams/domain/model/Cohort.java
@@ -1,8 +1,8 @@
 package fr.avenirsesr.portfolio.ams.domain.model;
 
 import fr.avenirsesr.portfolio.common.data.domain.model.AvenirsBaseModel;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.program.domain.model.TrainingPath;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 import java.time.Instant;
 import java.util.HashSet;
 import java.util.Set;
