+++ src/main/java/fr/avenirsesr/portfolio/file/domain/model/shared/File.java
@@ -1,7 +1,7 @@
 package fr.avenirsesr.portfolio.file.domain.model.shared;
 
 import fr.avenirsesr.portfolio.common.data.domain.model.AvenirsBaseModel;
-import fr.avenirsesr.portfolio.user.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import java.time.Instant;
 import java.util.UUID;
 import lombok.Getter;
