+++ src/main/java/fr/avenirsesr/portfolio/file/domain/port/input/UserResourceService.java
@@ -1,10 +1,10 @@
 package fr.avenirsesr.portfolio.file.domain.port.input;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.file.domain.model.EUserPhotoType;
 import fr.avenirsesr.portfolio.file.domain.model.UserPhoto;
 import fr.avenirsesr.portfolio.file.domain.model.UserPhotoUrlAndId;
-import fr.avenirsesr.portfolio.user.domain.model.User;
-import fr.avenirsesr.portfolio.user.domain.model.enums.EUserCategory;
 import java.io.IOException;
 import java.util.UUID;
