+++ src/main/java/fr/avenirsesr/portfolio/file/domain/port/output/repository/UserPhotoRepository.java
@@ -1,10 +1,10 @@
 package fr.avenirsesr.portfolio.file.domain.port.output.repository;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.common.data.domain.port.output.repository.GenericRepositoryPort;
 import fr.avenirsesr.portfolio.file.domain.model.EUserPhotoType;
 import fr.avenirsesr.portfolio.file.domain.model.UserPhoto;
-import fr.avenirsesr.portfolio.user.domain.model.User;
-import fr.avenirsesr.portfolio.user.domain.model.enums.EUserCategory;
 import java.util.List;
 import java.util.Optional;
