+++ src/main/java/fr/avenirsesr/portfolio/shared/application/adapter/utils/UserUtil.java
@@ -1,7 +1,7 @@
 package fr.avenirsesr.portfolio.shared.application.adapter.utils;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.user.domain.model.Student;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 import fr.avenirsesr.portfolio.user.domain.port.input.StudentService;
 import fr.avenirsesr.portfolio.user.domain.port.input.UserService;
 import java.security.Principal;
