+++ src/main/java/fr/avenirsesr/portfolio/trace/domain/port/input/TraceService.java
@@ -3,12 +3,12 @@ package fr.avenirsesr.portfolio.trace.domain.port.input;
 import fr.avenirsesr.portfolio.common.data.domain.model.DateFilter;
 import fr.avenirsesr.portfolio.common.data.domain.model.PageCriteria;
 import fr.avenirsesr.portfolio.common.data.domain.model.PagedResult;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.common.language.domain.model.enums.ELanguage;
 import fr.avenirsesr.portfolio.trace.domain.model.Trace;
 import fr.avenirsesr.portfolio.trace.domain.model.TraceDetail;
 import fr.avenirsesr.portfolio.trace.domain.model.TraceFilter;
 import fr.avenirsesr.portfolio.trace.domain.model.TracesSummary;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
