+++ src/main/java/fr/avenirsesr/portfolio/trace/domain/port/output/repository/TraceRepository.java
@@ -4,11 +4,11 @@ import fr.avenirsesr.portfolio.ams.domain.model.AMS;
 import fr.avenirsesr.portfolio.common.data.domain.model.DateFilter;
 import fr.avenirsesr.portfolio.common.data.domain.model.PageCriteria;
 import fr.avenirsesr.portfolio.common.data.domain.model.PagedResult;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.common.data.domain.port.output.repository.GenericDeletableRepositoryPort;
 import fr.avenirsesr.portfolio.student.progress.domain.model.SkillLevelProgress;
 import fr.avenirsesr.portfolio.trace.domain.model.Trace;
 import fr.avenirsesr.portfolio.trace.domain.model.TraceFilter;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 import java.util.List;
 
 public interface TraceRepository extends GenericDeletableRepositoryPort<Trace> {
