--- src/main/java/fr/avenirsesr/portfolio/user/domain/model/enums/EUserCategory.java
@@ -1,10 +1,0 @@
-package fr.avenirsesr.portfolio.user.domain.model.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum EUserCategory {
-  TEACHER,
-  STUDENT,
-  ;
-}
