+++ src/main/java/fr/avenirsesr/portfolio/user/domain/port/input/TeacherService.java
@@ -1,6 +1,6 @@
 package fr.avenirsesr.portfolio.user.domain.port.input;
 
-import fr.avenirsesr.portfolio.user.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 
 public interface TeacherService {
   String getBio(User user);
