+++ src/main/java/fr/avenirsesr/portfolio/user/domain/port/input/UserService.java
@@ -1,9 +1,9 @@
 package fr.avenirsesr.portfolio.user.domain.port.input;
 
-import fr.avenirsesr.portfolio.user.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.user.domain.model.UserPhotos;
 import fr.avenirsesr.portfolio.user.domain.model.UserProfileOverviewDTO;
-import fr.avenirsesr.portfolio.user.domain.model.enums.EUserCategory;
 import java.util.UUID;
 
 public interface UserService {
