+++ src/main/java/fr/avenirsesr/portfolio/user/domain/port/output/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package fr.avenirsesr.portfolio.user.domain.port.output.repository;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.common.data.domain.port.output.repository.GenericRepositoryPort;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 
 public interface UserRepository extends GenericRepositoryPort<User> {
   long countAll();
