+++ src/main/java/fr/avenirsesr/portfolio/user/domain/service/StudentServiceImpl.java
@@ -1,8 +1,8 @@
 package fr.avenirsesr.portfolio.user.domain.service;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.user.domain.exception.UserIsNotStudentException;
 import fr.avenirsesr.portfolio.user.domain.model.Student;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 import fr.avenirsesr.portfolio.user.domain.port.input.StudentService;
 import fr.avenirsesr.portfolio.user.domain.port.output.repository.StudentRepository;
 import java.util.UUID;
