+++ src/main/java/fr/avenirsesr/portfolio/user/infrastructure/adapter/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package fr.avenirsesr.portfolio.user.infrastructure.adapter.mapper;
 
-import fr.avenirsesr.portfolio.user.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.user.infrastructure.adapter.model.UserEntity;
 
 public interface UserMapper {
