+++ src/main/java/fr/avenirsesr/portfolio/user/infrastructure/adapter/model/ExternalUserEntity.java
@@ -1,8 +1,8 @@
 package fr.avenirsesr.portfolio.user.infrastructure.adapter.model;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.common.data.infrastructure.adapter.model.AvenirsBaseEntity;
 import fr.avenirsesr.portfolio.user.domain.model.enums.EExternalSource;
-import fr.avenirsesr.portfolio.user.domain.model.enums.EUserCategory;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
