+++ src/main/java/fr/avenirsesr/portfolio/user/infrastructure/adapter/repository/UserDatabaseRepository.java
@@ -1,7 +1,7 @@
 package fr.avenirsesr.portfolio.user.infrastructure.adapter.repository;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.common.data.infrastructure.adapter.repository.GenericJpaRepositoryAdapter;
-import fr.avenirsesr.portfolio.user.domain.model.User;
 import fr.avenirsesr.portfolio.user.domain.port.output.repository.UserRepository;
 import fr.avenirsesr.portfolio.user.infrastructure.adapter.mapper.UserMapper;
 import fr.avenirsesr.portfolio.user.infrastructure.adapter.model.UserEntity;
