+++ src/main/java/fr/avenirsesr/portfolio/user/infrastructure/adapter/seeder/fake/FakeExternalUser.java
@@ -1,9 +1,9 @@
 package fr.avenirsesr.portfolio.user.infrastructure.adapter.seeder.fake;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.common.seeder.domain.port.output.SharedDataGenerator;
 import fr.avenirsesr.portfolio.common.seeder.infrastructure.adapter.data.DataGeneratorProvider;
 import fr.avenirsesr.portfolio.user.domain.model.enums.EExternalSource;
-import fr.avenirsesr.portfolio.user.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.user.domain.port.output.seeder.UserDataGenerator;
 import fr.avenirsesr.portfolio.user.infrastructure.adapter.model.ExternalUserEntity;
 import fr.avenirsesr.portfolio.user.infrastructure.adapter.model.UserEntity;
