+++ src/test/java/fr/avenirsesr/portfolio/file/infrastructure/fixture/UserPhotoFixture.java
@@ -1,9 +1,9 @@
 package fr.avenirsesr.portfolio.file.infrastructure.fixture;
 
+import fr.avenirsesr.portfolio.common.data.domain.model.User;
+import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
 import fr.avenirsesr.portfolio.file.domain.model.EUserPhotoType;
 import fr.avenirsesr.portfolio.file.domain.model.shared.EFileType;
-import fr.avenirsesr.portfolio.user.domain.model.User;
-import fr.avenirsesr.portfolio.user.domain.model.enums.EUserCategory;
 import java.time.Instant;
 import java.util.UUID;
