+++ src/test/java/fr/avenirsesr/portfolio/student/progress/application/adapter/controller/StudentProgressControllerIT.java
@@ -163,8 +163,7 @@ class StudentProgressControllerIT {
         .andExpect(jsonPath("$[0].skills[0].name").value("Skill culpa"))
         .andExpect(jsonPath("$[0].skills[0].levelCount").value(3))
         .andExpect(jsonPath("$[0].skills[0].currentSkillLevel").exists())
-        .andExpect(jsonPath("$[0].skills[0].currentSkillLevel.traceCount").value(0))
-        .andExpect(jsonPath("$[0].skills[0].currentSkillLevel.activityCount").value(0));
+        .andExpect(jsonPath("$[0].skills[0].currentSkillLevel.traceCount").value(0));
   }
 
   @Test
